+++ tests/testthat.R
@@ -1,5 +1,4 @@
 library('testthat')
 library('Ternary')
-library('vdiffr')
 
 test_check("Ternary")